fix: resolve Vercel deployment ENOENT error for route group pages
- Extract client components from (dashboard) route group pages into
  components/dashboard/ and use server component wrappers to avoid
  page_client-reference-manifest.js ENOENT error on Vercel
- Add output: 'standalone' to next.config.mjs for Vercel compatibility
- Remove experimental.serverComponentsExternalPackages (not needed)
- Remove postinstall prisma generate script (Vercel build is just next build)
